+++ Evaluation/src/metrics/single/boardCoverage/BoardCoverageDefault.java
@@ -74,7 +74,7 @@ public class BoardCoverageDefault extends Metric
 			numSitesCovered += ((double) sitesCovered.size()) / context.board().topology().getGraphElements(context.board().defaultSite()).size();
 		}
 
-		return numSitesCovered / trials.length;
+		return Double.valueOf(numSitesCovered / trials.length);
 	}
 
 }
